Added a check for an MPI compiler in OmicABELnoMM's configure.ac. Also added an option to so that the user can ask explicitly to use MPI (./configure --with-mpi) or explicitly disable use of MPI even if compilers are found (./configure --without-mpi) if neither the --with or --without option is specified, ./configure will try to locate an MPI compiler and if that fails, it will continue without.
Based on the results of this check, the variable USE_MPI is set to "yes" and passed to Automake for use during compilation and in the source code (usinf #ifdef).


NOTE for developers: The AX_PROG_CXX_MPI Autoconf macro is not part of the default Autoconf installation. It can be downloaded from the GNU Autoconf Archive: http://www.gnu.org/software/autoconf-archive/ or installed via the package manager of most Linux distributions, e.g. in Debian and its derivatives the package is called 'autoconf-archive'.


git-svn-id: svn+ssh://svn.r-forge.r-project.org/svnroot/genabel/pkg/OmicABELnoMM@1849 edb9625f-4e0d-4859-8d74-9fd3b1da38cb