Add configure option to create the data files
With the move to GitHub, we had to remove the example data files because
they were over 100MB. As a result, the make step failed because there is
no Makefile entry that generates the data.

This adds a user configurable option (--enable-examples) to ./configure
that the user must explicitly mention when (s)he wants to build the
large example files (running the R script takes a while).

Note that the large example files will be removed when running 'make clean'.