Fix 'make check' and large example files
With this commit the 'make check' command works again. The Makefile.am's
in the examples and tests directories are now set up in such a way that
the large example files are not distributed in the tar.gz file, but will
be generated once the user runs 'make check'.
(Like before, the smaller .fv? files are distributed in the tar.gz
file).